requirements.txt
setuptools
uvicorn
gunicorn
connexion[swagger-ui]
a2wsgi
networkx
python-dotenv